+++ ml_peg/app/build_app.py
@@ -3,15 +3,12 @@
 from __future__ import annotations
 
 from importlib import import_module
-import random
 import warnings
 
 from dash import (
-    ALL,
     Dash,
     Input,
     Output,
-    State,
     callback,
     clientside_callback,
     ctx,
@@ -20,7 +17,7 @@ from dash import (
 from dash.dash_table import DataTable
 from dash.dcc import Dropdown, Interval, Link, Loading, Location, Store
 from dash.exceptions import PreventUpdate
-from dash.html import H1, H3, A, Br, Button, Details, Div, Img, Span, Summary
+from dash.html import H1, H3, A, Br, Details, Div, Img, Span, Summary
 from yaml import safe_load
 
 from ml_peg.analysis.utils.utils import calc_table_scores, get_table_style
@@ -56,14 +53,15 @@ from ml_peg.app.utils.utils import (
     load_model_registry_configs,
     sig_fig_format,
 )
+from ml_peg.app.utils.weight_presets import (
+    build_weight_preset_selector,
+    register_weight_preset_callbacks,
+)
 from ml_peg.models import current_models
 from ml_peg.models.get_models import get_model_names
 
 # Get all models
 MODELS = get_model_names(current_models)
-
-# Summary-table category-weight presets (relevant categories -> weight 1, rest 0).
-WEIGHT_PRESETS = safe_load((APP_ROOT / "utils" / "weight_presets.yml").read_text())
 
 
 def _nav_link_style(is_active: bool) -> dict[str, str]:
@@ -89,31 +87,6 @@ def _nav_link_style(is_active: bool) -> dict[str, str]:
         "fontWeight": "600" if is_active else "normal",
         "backgroundColor": "#e8f4ff" if is_active else "transparent",
         "borderLeft": ("3px solid #119DFF" if is_active else "3px solid transparent"),
-    }
-
-
-def _preset_btn_style(selected: bool) -> dict[str, str]:
-    """
-    Return the style for a weight-preset toggle button.
-
-    Parameters
-    ----------
-    selected
-        Whether the preset is currently selected.
-
-    Returns
-    -------
-    dict[str, str]
-        Style dictionary for the button.
-    """
-    return {
-        "padding": "4px 10px",
-        "fontSize": "12px",
-        "borderRadius": "4px",
-        "cursor": "pointer",
-        "border": "1px solid #228be6" if selected else "1px solid #ced4da",
-        "backgroundColor": "#228be6" if selected else "#fff",
-        "color": "#fff" if selected else "#495057",
     }
 
 
@@ -998,64 +971,7 @@ def build_nav(
         style={"marginBottom": "8px", "fontSize": "13px"},
     )
 
-    preset_names = [*WEIGHT_PRESETS, "Random"]
-    weight_preset_selector = Details(
-        [
-            Summary("Weight presets", style=_summary_label_style),
-            Div(
-                [
-                    Div(
-                        [
-                            Button(
-                                name,
-                                id={"type": "weight-preset-btn", "index": name},
-                                n_clicks=0,
-                                style=_preset_btn_style(False),
-                            )
-                            for name in preset_names
-                        ],
-                        style={"display": "flex", "flexWrap": "wrap", "gap": "6px"},
-                    ),
-                    Div(
-                        [
-                            Button(
-                                "Apply",
-                                id="weight-preset-apply",
-                                n_clicks=0,
-                                style={
-                                    "padding": "4px 12px",
-                                    "fontSize": "12px",
-                                    "backgroundColor": "#228be6",
-                                    "color": "#fff",
-                                    "border": "none",
-                                    "borderRadius": "4px",
-                                    "cursor": "pointer",
-                                },
-                            ),
-                            Button(
-                                "Reset",
-                                id="weight-preset-reset",
-                                n_clicks=0,
-                                style={
-                                    "padding": "4px 12px",
-                                    "fontSize": "12px",
-                                    "backgroundColor": "#fff",
-                                    "color": "#495057",
-                                    "border": "1px solid #ced4da",
-                                    "borderRadius": "4px",
-                                    "cursor": "pointer",
-                                },
-                            ),
-                        ],
-                        style={"display": "flex", "gap": "6px", "marginTop": "8px"},
-                    ),
-                    Store(id="weight-preset-pending", storage_type="memory", data=[]),
-                ],
-                style={"padding": "8px 12px"},
-            ),
-        ],
-        style={"marginBottom": "8px", "fontSize": "13px"},
-    )
+    weight_preset_selector = build_weight_preset_selector(_summary_label_style)
 
     sidebar = Div(
         id="sidebar-nav",
@@ -1357,142 +1273,9 @@ def build_nav(
             return selected, selected
         raise PreventUpdate
 
-    @callback(
-        Output("weight-preset-pending", "data", allow_duplicate=True),
-        Input({"type": "weight-preset-btn", "index": ALL}, "n_clicks"),
-        State("weight-preset-pending", "data"),
-        prevent_initial_call=True,
+    register_weight_preset_callbacks(
+        summary_table, _default_weight_store_data(summary_table)
     )
-    def toggle_weight_preset(_n_clicks: list[int], pending: list[str]) -> list[str]:
-        """
-        Toggle the clicked preset in the pending selection.
-
-        Parameters
-        ----------
-        _n_clicks
-            Click counts for all preset buttons.
-        pending
-            Currently pending preset selection.
-
-        Returns
-        -------
-        list[str]
-            Updated pending preset selection.
-        """
-        trigger = ctx.triggered_id
-        if not isinstance(trigger, dict):
-            raise PreventUpdate
-        name = trigger["index"]
-        pending = list(pending or [])
-        if name in pending:
-            pending.remove(name)
-        else:
-            pending.append(name)
-        return pending
-
-    @callback(
-        Output({"type": "weight-preset-btn", "index": ALL}, "style"),
-        Input("weight-preset-pending", "data"),
-        State({"type": "weight-preset-btn", "index": ALL}, "id"),
-        prevent_initial_call=True,
-    )
-    def sync_weight_preset_styles(
-        pending: list[str], ids: list[dict]
-    ) -> list[dict[str, str]]:
-        """
-        Highlight preset buttons that are currently selected.
-
-        Parameters
-        ----------
-        pending
-            Currently pending preset selection.
-        ids
-            Ids of the preset buttons, in layout order.
-
-        Returns
-        -------
-        list[dict[str, str]]
-            Style dictionaries for each preset button.
-        """
-        selected = set(pending or [])
-        return [_preset_btn_style(button["index"] in selected) for button in ids]
-
-    @callback(
-        Output("summary-table-weight-store", "data", allow_duplicate=True),
-        Input("weight-preset-apply", "n_clicks"),
-        State("weight-preset-pending", "data"),
-        prevent_initial_call=True,
-    )
-    def apply_weight_presets(n_clicks: int, pending: list[str]) -> dict[str, float]:
-        """
-        Apply the union of the selected presets to the summary weights.
-
-        A category is weighted 1.0 if it appears in any selected preset, else 0.0
-        (the union of the selected presets). "Random" instead gives every category
-        an independent random weight in [0, 1].
-
-        Parameters
-        ----------
-        n_clicks
-            Apply-button click count.
-        pending
-            Currently selected presets.
-
-        Returns
-        -------
-        dict[str, float]
-            Category-column weights written to the summary weight store.
-        """
-        if not n_clicks or not pending:
-            raise PreventUpdate
-
-        weight_columns = [
-            column["id"]
-            for column in summary_table.columns
-            if column["id"] not in {"MLIP", "Score", "id", "link"}
-        ]
-
-        weights = dict.fromkeys(weight_columns, 0.0)
-        for name in pending:
-            if name == "Random":
-                contribution = {
-                    column: round(random.random(), 2) for column in weight_columns
-                }
-            else:
-                relevant = {f"{title} Score" for title in WEIGHT_PRESETS[name]}
-                contribution = {
-                    column: (1.0 if column in relevant else 0.0)
-                    for column in weight_columns
-                }
-            weights = {
-                column: max(weights[column], contribution[column])
-                for column in weight_columns
-            }
-        return weights
-
-    @callback(
-        Output("summary-table-weight-store", "data", allow_duplicate=True),
-        Output("weight-preset-pending", "data", allow_duplicate=True),
-        Input("weight-preset-reset", "n_clicks"),
-        prevent_initial_call=True,
-    )
-    def reset_weight_presets(n_clicks: int) -> tuple[dict[str, float], list[str]]:
-        """
-        Restore the category yml default weights and clear the selection.
-
-        Parameters
-        ----------
-        n_clicks
-            Reset-button click count.
-
-        Returns
-        -------
-        tuple[dict[str, float], list[str]]
-            Default category weights and an empty pending selection.
-        """
-        if not n_clicks:
-            raise PreventUpdate
-        return _default_weight_store_data(summary_table), []
 
     @callback(
         Output("model-filter-details", "open"),
